+++ Codoxide.Outcome.Extensions.Tap/src/Tap.extensions.cs
